requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.4.0
aiosmtplib==4.0.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==24.2.0
black==25.1.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
coverage==7.2.7
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
duckdb==1.3.0
duckduckgo_search==8.0.4
email_validator==2.2.0
executing==2.1.0
faiss-cpu==1.11.0
fastapi==0.111.0
fastapi-cli==0.0.8
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
frappe-bench==5.23.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
honcho==2.0.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.68
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.3.45
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.17.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
ollama==0.6.0
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.6.0
postgrest==1.0.2
preshed==3.0.10
primp==0.15.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.0
pytest-asyncio==0.23.8
pytest-cov==4.1.0
pytest-mock==3.11.1
python-crontab==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.4.3
redis==5.0.4
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.8
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.0
scipy==1.16.0
semantic-version==2.10.0
sentence-transformers==5.1.0
setuptools==78.0.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.37.2
storage3==0.11.3
streamlit==1.41.0
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
tenacity==9.0.0
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
ujson==5.10.0
unittest-xml-reporting==3.2.0
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.30.1
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==14.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0